+++ dbms/src/AggregateFunctions/tests/gtest_agg_func_return_type.cpp
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <AggregateFunctions/AggregateFunctionFactory.h>
 #include <AggregateFunctions/AggregateFunctionSum.h>
 #include <DataTypes/DataTypeDecimal.h>
 #include <DataTypes/DataTypeNullable.h>
@@ -31,30 +30,32 @@ namespace tests
 class ExecutorAggFuncReturnTypeTestRunner : public DB::tests::AggregationTest
 {
 public:
-    ::testing::AssertionResult checkAggReturnType(const String & agg_name, const DataTypes & data_types, const DataTypePtr & expect_type)
-    {
-        AggregateFunctionPtr agg_ptr = DB::AggregateFunctionFactory::instance().get(agg_name, data_types, {});
-        const DataTypePtr & ret_type = agg_ptr->getReturnType();
-        if (ret_type->getName() == expect_type->getName())
-            return ::testing::AssertionSuccess();
-        return ::testing::AssertionFailure() << "Expect type: " << expect_type->getName() << " Actual type: " << ret_type->getName();
-    }
-
     template <typename T>
     ::testing::AssertionResult testSumOnPartialResult()
     {
         for (size_t nullable = 0; nullable <= 1; ++nullable)
         {
-            PrecType precision = maxDecimalPrecision<T>() - 1;
-            ScaleType scale = 5;
-            std::shared_ptr<DataTypeDecimal<T>> input_nested_type_ptr = std::make_shared<DataTypeDecimal<T>>(precision, scale);
-            std::shared_ptr<DataTypeDecimal<T>> expect_output_nested_type_ptr = std::make_shared<DataTypeDecimal<T>>(precision, scale);
-            DataTypePtr input_type_ptr = nullable == 0 ? static_cast<DataTypePtr>(input_nested_type_ptr) : static_cast<DataTypePtr>(std::make_shared<DataTypeNullable>(input_nested_type_ptr));
-            DataTypePtr expect_output_type_ptr = nullable == 0 ? static_cast<DataTypePtr>(expect_output_nested_type_ptr) : static_cast<DataTypePtr>(std::make_shared<DataTypeNullable>(expect_output_type_ptr));
-            auto result = checkAggReturnType(NameSumOnPartialResult::name, {input_type_ptr}, expect_output_type_ptr);
-            if (result)
-                continue;
-            return result;
+            PrecType max_precision = maxDecimalPrecision<T>();
+            std::vector<std::pair<PrecType, ScaleType>> cases{
+                {max_precision, 0},
+                {max_precision-1, 1},
+                {max_precision-2, 2},
+                {max_precision-3, 3},
+            };
+
+            for (const auto & p : cases)
+            {
+                PrecType precision = p.first;
+                ScaleType scale = p.second;
+                std::shared_ptr<DataTypeDecimal<T>> input_nested_type_ptr = std::make_shared<DataTypeDecimal<T>>(precision, scale);
+                std::shared_ptr<DataTypeDecimal<T>> expect_output_nested_type_ptr = std::make_shared<DataTypeDecimal<T>>(precision, scale);
+                DataTypePtr input_type_ptr = nullable == 0 ? static_cast<DataTypePtr>(input_nested_type_ptr) : static_cast<DataTypePtr>(std::make_shared<DataTypeNullable>(input_nested_type_ptr));
+                DataTypePtr expect_output_type_ptr = nullable == 0 ? static_cast<DataTypePtr>(expect_output_nested_type_ptr) : static_cast<DataTypePtr>(std::make_shared<DataTypeNullable>(expect_output_nested_type_ptr));
+                auto result = checkAggReturnType(NameSumOnPartialResult::name, {input_type_ptr}, expect_output_type_ptr);
+                if (result)
+                    continue;
+                return result;
+            }
         }
 
         return ::testing::AssertionSuccess();
@@ -64,7 +65,7 @@ public:
 TEST_F(ExecutorAggFuncReturnTypeTestRunner, AggregationSum)
 try
 {
-    // Test output type of the second stage sum aggregation function
+    // Test output type of the sum aggregation function which receives partial result
     ASSERT_TRUE(testSumOnPartialResult<Decimal32>());
     ASSERT_TRUE(testSumOnPartialResult<Decimal64>());
     ASSERT_TRUE(testSumOnPartialResult<Decimal128>());
