+++ dbms/src/TestUtils/AggregationTestUtils.h
@@ -17,12 +17,23 @@
 #include <TestUtils/TiFlashTestEnv.h>
 #include <AggregateFunctions/registerAggregateFunctions.h>
 #include <gtest/gtest.h>
+#include <AggregateFunctions/IAggregateFunction.h>
+#include <AggregateFunctions/AggregateFunctionFactory.h>
 
 namespace DB::tests
 {
 
 class AggregationTest : public ::testing::Test {
 public:
+    ::testing::AssertionResult checkAggReturnType(const String & agg_name, const DataTypes & data_types, const DataTypePtr & expect_type)
+    {
+        AggregateFunctionPtr agg_ptr = DB::AggregateFunctionFactory::instance().get(agg_name, data_types, {});
+        const DataTypePtr & ret_type = agg_ptr->getReturnType();
+        if (ret_type->getName() == expect_type->getName())
+            return ::testing::AssertionSuccess();
+        return ::testing::AssertionFailure() << "Expect type: " << expect_type->getName() << " Actual type: " << ret_type->getName();
+    }
+
     static void SetUpTestCase();
 };
